requirements.txt
RPi.GPIO==0.7.1
opencv-python==4.8.1.78
imutils==0.5.4
Adafruit_DHT==1.4.0
